test(sse): split the terminal-marker suite out of stream-handler tests
fork.7 grew `stream-handler.test.ts` to 873 lines, past the 800-line cap
`check:test-file-size` enforces for files that are not in the frozen baseline,
so the gate has been red on `stable` since that commit.

Move the six tests that cover "a stream which already delivered its terminal SSE
event must not be finalized as a client disconnect" into their own file. Pure
move plus the imports each half still needs: 665 + 237 lines, the same 27 tests,
all passing.
